Centralise the data types and correct the imports
