+++ frontend/src/types/event.ts
@@ -1,3 +1,5 @@
+export type ProfileTab = 'created' | 'going' | 'archived';
+
 export type Attendee = {
   id: string;
   firstName: string;
@@ -24,4 +26,44 @@ export type EventData = {
   attendeeCount: number;
   attendees: string[];
   attendances: Attendance[];
+};
+
+export type RawProfileEvent = {
+  id: string;
+  title: string;
+  description: string | null;
+  date: string;
+  location: string;
+  capacity: number;
+  organizerId: string;
+  organizer: { firstName: string; lastName: string };
+  _count: { attendances: number };
+};
+
+export type ProfileEvent = {
+  id: string;
+  status: ProfileTab;
+  date: Date;
+  title: string;
+  author: string;
+  description: string;
+  attendeeCount: number;
+  maxAttendees: number;
+};
+
+export type RawEvent = {
+  id: string;
+  title: string;
+  description: string;
+  date: string;
+  location: string;
+  capacity: number;
+  organizer: {
+    id: string;
+    firstName: string;
+    lastName: string;
+  };
+  attendances: Attendance[];
+  isOwner: boolean;
+  isAttending: boolean;
 };
